move templates into separate class
