+++ analyzer/src/main/groovy/com/novoda/sqlite/generator/SqliteAnalyzerPlugin.groovy
@@ -14,7 +14,6 @@ class SqliteAnalyzerPlugin implements Plugin<Project> {
                         migrationsDir project.file(extension.migrationsDir)
                         outputDir sourceFolder
                         packageName extension.packageName
-                        generateAuto extension.generateAuto
                         onlyStatic extension.onlyStatic
                     }
                     variant.registerJavaGeneratingTask(javaGenerationTask, sourceFolder)
@@ -27,7 +26,6 @@ class SqliteAnalyzerPlugin implements Plugin<Project> {
                         databaseFile project.file(extension.databaseFile)
                         outputDir sourceFolder
                         packageName extension.packageName
-                        generateAuto extension.generateAuto
                         onlyStatic extension.onlyStatic
                     }
                     variant.registerJavaGeneratingTask(javaGenerationTask, sourceFolder)
